chore: separate output and input entities
