+++ packages/api/src/tag/dto/list-tags.output.ts
@@ -1,0 +1,9 @@
+import { Field, ObjectType } from '@nestjs/graphql'
+import { Output } from 'src/types/output/output.entity'
+import { Tag } from '../entities/tag.entity'
+
+@ObjectType()
+export class ListTagsOutput extends Output {
+  @Field(() => [Tag])
+  tags?: Tag[]
+}
