+++ packages/api/src/todo-group/dto/create-todo-group.input.ts
@@ -1,11 +1,4 @@
-import {
-  Field,
-  InputType,
-  ObjectType,
-  OmitType,
-  PartialType,
-} from '@nestjs/graphql'
-import { Output } from 'src/types/output/output.entity'
+import { Field, InputType, OmitType, PartialType } from '@nestjs/graphql'
 import { TodoGroup } from '../entities/todo_group.entity'
 
 @InputType()
@@ -14,10 +7,4 @@ export class CreateTodoGroupInput extends OmitType(PartialType(TodoGroup), [
 ]) {
   @Field(() => String)
   name: string
-}
-
-@ObjectType()
-export class CreateTodoGroupOutput extends Output {
-  @Field(() => TodoGroup)
-  todoGroup?: TodoGroup
 }
