+++ packages/api/src/user/dto/create-user.input.ts
@@ -1,6 +1,4 @@
-import { Field, InputType, ObjectType, PickType } from '@nestjs/graphql'
-import { Output } from 'src/types/output/output.entity'
-
+import { Field, InputType, PickType } from '@nestjs/graphql'
 import { User } from '../entities/user.entity'
 
 @InputType()
@@ -13,10 +11,4 @@ export class CreateUserInput extends PickType(User, [
 
   @Field({ description: 'user password' })
   password: string
-}
-
-@ObjectType()
-export class CreateUserOutput extends Output {
-  @Field(() => User, { nullable: true })
-  user?: User
 }
