+++ packages/api/src/user/dto/update-user.input.ts
@@ -1,5 +1,4 @@
 import { Field, InputType, PartialType } from '@nestjs/graphql'
-
 import { CreateUserInput } from './create-user.input'
 
 @InputType()
